refactor: searchBar to context
