+++ src/components/SearchTermContext.types.ts
@@ -1,0 +1,8 @@
+export interface SearchTermProviderProps {
+  children: React.ReactNode;
+}
+
+export interface SearchTermContextType {
+  searchTerm: string;
+  saveSearchData: (term: string) => void;
+}
